+++ index.html
@@ -385,6 +385,6 @@
       >
     </footer>
 
-    <script src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
   </body>
 </html>
