+++ js/typeFilter.js
@@ -1,0 +1,27 @@
+export default function typeFilter() {
+  const filters = document.querySelectorAll('[data-filter]');
+  filters[0].classList.add('active');
+
+  function handleFilter(e, index) {
+    const filter = e.target.dataset.filter;
+    filters.forEach((filter) => filter.classList.remove('active'));
+    filters[index].classList.add('active');
+    filterType(filter);
+  }
+
+  function filterType(type) {
+    if (type === 'all') return false;
+    const typeItens = [...document.querySelectorAll('[data-type]')];
+    typeItens.forEach((item) => item.classList.add('hide'));
+    const filteredItens = typeItens.filter(
+      (item) => item.dataset.type === type,
+    );
+    filteredItens.forEach((item) => item.classList.remove('hide'));
+  }
+
+  filters.forEach((filter, index) => {
+    filter.addEventListener('click', (e) => handleFilter(e, index));
+  });
+
+  filterType('all');
+}
